Fix dropped packets in mpeg stream
In mpeg_bitstream_parse, packet->status can be set to
LIBCAPTION_READY or LIBCAPTION_ERROR before we have parsed all of the data
within the bitstream buffer belonging to the packet.  If this happens,
then we were skipping over the remaining chunk the next time this
function was called.

Add a parse marker variable to the packet to keep track of our last
parsing position.