h264 avc/mp4: add support for parsing NALU
The h264 avc/mp4 format does not have a start code of 00 00 01, but
instead contains a fixed length NALU length field.  Knowledge of whether
or not to use this mode needs to be provided by the consumer of the
libcaption API, so add a new function to allow for this format.